requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cmake==3.29.3
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.3
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jupyter-client==8.6.2
jupyter-core==5.7.2
jupyterlab-widgets==3.0.11
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
lit==18.1.6
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt-toolkit==3.0.46
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.0.0
typing-extensions==4.5.0
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
urllib3==2.2.1
wcwidth==0.2.13
werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
zipp==3.19.2
